feat: optimize topics page
