+++ apps/the_monkeys/src/app/layout.tsx
@@ -1,10 +1,10 @@
 import type { Metadata } from 'next';
+import dynamic from 'next/dynamic';
 import { DM_Sans, Inter, Newsreader } from 'next/font/google';
 import Script from 'next/script';
 
 import { AppShell } from '@/components/layout/app-shell/AppShell';
 import Footer from '@/components/layout/footer';
-import DonationPopup from '@/components/popup/DonationPopup';
 import { LIVE_URL } from '@/constants/api';
 import { TooltipProvider } from '@the-monkeys/ui/atoms/tooltip';
 import { Toaster } from '@the-monkeys/ui/molecules/toaster';
@@ -14,6 +14,8 @@ import './globals.css';
 import GrowthbookClientProvider from './growthbook-provider';
 import { QueryClientMount } from './query-client-mount';
 import { ThemeProviders } from './theme-provider';
+
+const DonationPopup = dynamic(() => import('@/components/popup/DonationPopup'));
 
 const inter = Inter({
   weight: ['300', '400', '500', '600', '700'],
